+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Refactored referral param in advanced search (#326)
 * Refactored the process of check permission
 * Refactored the process of get_available_attrs in Entry
+* Refactored drf response format and default settings
 
 ## v3.4.1